fill testing

+++ primitives.h
@@ -11,9 +11,9 @@
 struct polygon
 {
 	int x, y;
-	double angle;
-	double scale_x;
-	double scale_y;
+	float angle;
+	float scale_x;
+	float scale_y;
 	int nsides;
 	SDL_Point *vectors;
 	SDL_Point *points;
@@ -22,9 +22,9 @@ struct polygon
 struct ellipse
 {
 	int x, y;
-	double angle;
-	double scale_x;
-	double scale_y;
+	float angle;
+	float scale_x;
+	float scale_y;
 	int radius_x;
 	int radius_y;
 };
@@ -38,14 +38,14 @@ int draw_ellipse								(SDL_Renderer *renderer, const struct ellipse *e);
 int draw_ellipse_filled						(SDL_Renderer *renderer, const struct ellipse *e);
 
 // build polygons
-struct polygon *create_polygon			(SDL_Point *vectors, int nsides, int x, int y, double angle, double scale_x, double scale_y);
-struct polygon *create_reg_polygon		(int nsides, int x, int y, double radius, double angle, double scale_x, double scale_y);
-struct polygon *create_random_polygon	(int nsides, int x, int y, double max_radius, double min_radius, double angle, double scale_x, double scale_y);
+struct polygon *create_polygon			(SDL_Point *vectors, int nsides, int x, int y, float angle, float scale_x, float scale_y);
+struct polygon *create_reg_polygon		(int nsides, int x, int y, float radius, float angle, float scale_x, float scale_y);
+struct polygon *create_random_polygon	(int nsides, int x, int y, float max_radius, float min_radius, float angle, float scale_x, float scale_y);
 
 // transform polygon
 int polygon_translate						(struct polygon *p, int x, int y);
-int polygon_angle								(struct polygon *p, double angle);
-int polygon_scale								(struct polygon *p, double scale_x, double scale_y);
+int polygon_angle								(struct polygon *p, float angle);
+int polygon_scale								(struct polygon *p, float scale_x, float scale_y);
 int polygon_rebuild							(struct polygon *p);
 
 // free polygon
